fixed delete-user route

+++ client/src/components/DeleteProfileButton/DeleteProfileButton.jsx
@@ -1,46 +1,44 @@
-import React, { useContext } from "react";
-import axios from "axios";
-import "./DeleteProfileButton.css";
-import globalUser from "../../utils/globalUser";
-import { useParams, useHistory} from "react-router-dom";
 
-
-function DeleteProfileButton() {
+import React, { useContext } from 'react'
+import axios from 'axios'
+import './DeleteProfileButton.css'
+import globalUser from '../../utils/globalUser'
+import { useParams, useHistory, Link } from 'react-router-dom'
+function DeleteProfileButton () {
   // const { id } = useParams();
   const id = localStorage.getItem('user')
-  // const history = useHistory();
-
+  const history = useHistory()
   const deleteProfile = () => {
-    axios
-      .delete(`/delete-user/${id}`)
-      .then(() => {
-        alert("profile successfully deleted!");
-        localStorage.clear();
-        // history.push("/welcome");
-        console.log("profile deleted");
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-    
-  };
-  
-  
+    if (id) {
+      axios
+        .delete(`/api/users/delete-user/${id}`)
+        .then(() => {
+          alert('profile successfully deleted!')
+          localStorage.clear()
+          // history.push("/welcome");
+          console.log('profile deleted')
+        })
+        .catch(err => {
+          console.log(err)
+        })
+    }
+  }
   return (
     <div>
-      <button
-        className="btn hoverable"
-        id="delete-button"
-        to = "/welcome"
-        onClick={() => {
-          deleteProfile();
-        }}
-       
-      >
-        Delete Insulterator9000 Profile
-    </button>
+      <Link to='/'>
+        <button
+          className='btn hoverable'
+          id='delete-button'
+          // to="/welcome"
+          onClick={() => {
+            deleteProfile()
+            // history.push("/welcome");
+          }}
+        >
+          Delete Insulterator9000 Profile
+        </button>
+      </Link>
     </div>
-  );
-};
-
-export default DeleteProfileButton;
+  )
+}
+export default DeleteProfileButton
